Bug Fix: real -> int this screwed up windowing

+++ src/mc/MC_move.f90
@@ -76,7 +76,7 @@ function exponential_random_int(window,rand_stat) result(output)
     implicit none
     type(random_stat), intent(inout) :: rand_stat  ! status of random number generator
     real urnd(1) ! single random number
-    integer, intent(in) :: window
+    real(dp), intent(in) :: window
     integer output
     call random_number(urnd,rand_stat)
     output  = nint(-1.0_dp*log(urnd(1)+0.000001_dp)*window+0.0001_dp)
